backend/requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5
Pillow>=9.0
django-import-export>=3.0
django-filter>=22.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-widget-tweaks>=1.4
python-dateutil>=2.8
gunicorn>=20.1.0
requests>=2.25.0
pytest>=6.0
pytest-django>=4.0
django-simple-history>=3.5
django-redis>=5.4
psycopg2-binary>=2.9
openpyxl>=3.1
reportlab>=4.4
dj-database-url>=2.0
celery>=5.3
redis>=5.0
django-celery-beat>=2.6
django-cors-headers>=4.3
python-dotenv>=1.0.0
whitenoise>=6.6.0
